RDKCOM-2792: REFPLTB-1383,REFPLTB-1461 : Need to resolve build time warning ...
Reason for change: Adding 64bit support for rpi
Test Procedure: Built 32 bit and 64 bit image
Risks: Low

(cherry picked from commit 58c35305e282debfd3753debffd6a2c505ae222c)

+++ source-embedded/CcspCwmpProcessor/ccsp_cwmp_proco_acs_access.c
@@ -456,7 +456,7 @@ void
 CcspCwmppoMsValueChanged
     (
         ANSC_HANDLE                 hThisObject,
-        ULONG                       ulMsvcObjectID
+        int                         ulMsvcObjectID
     )
 {
     PCCSP_CWMP_PROCESSOR_OBJECT     pMyObject           = (PCCSP_CWMP_PROCESSOR_OBJECT )hThisObject;
